+++ ml_tools/keys/_keys.py
@@ -100,7 +100,7 @@ class CaptumKeys:
     IMPORTANCE_COLUMN = "Scaled Mean Attribution"
     PERCENT_COLUMN = "Relative Importance(%)"
     SAVENAME = "captum_summary"
-    PLOT_NAME = "captum_importance_plot"
+    PLOT_NAME = "captum_importance"
 
 
 class PyTorchCheckpointKeys:
@@ -261,6 +261,10 @@ class _EvaluationConfig:
     REGRESSION_PLOT_SIZE = (10, 7)
     SEQUENCE_PLOT_SIZE = (10, 7)
     CLASSIFICATION_PLOT_SIZE = (9, 9)
+    # Captum plots
+    CAPTUM_PLOT_SIZE = (10, 8)
+    CAPTUM_FONT_SIZE = 24
+    CAPTUM_X_TICK_SIZE = 20
     # Loss plot
     LOSS_PLOT_SIZE = (18, 9)
     LOSS_PLOT_LABEL_SIZE = 24
@@ -270,6 +274,7 @@ class _EvaluationConfig:
     CM_SIZE = (9, 8)    # used for multi label binary classification confusion matrix 
     HEATMAP_WIDTH = 10.0 # default width for classification heatmaps, height is dynamic based on number of classes
     NAME_LIMIT = 15  # max number of characters for feature/label names in plots
+    TITLE_LIMIT = 35  # max number of characters for plot titles before wrapping
     # RADAR Plot settings
     RADAR_PLOT_WIDTH = 800
     RADAR_PLOT_HEIGHT = 800